Code refactoring (#340)
* :books: add migration notes. #322

* :tractor: relocate files for coherence

* :tractor: relocated plugins.py

* :tractor: grouped all plugins to plugins directory

* :tractor: relocated data loader manager to core folder

* :tractor: relocated definitions to core folder

* :tractor: relocated hashstore to core folder

* :green_heart: enable plugins to be found as module

* :shirt: update coding style

* :green_heart: update hardcode references

* :tractor: relocate utils
