+++ docs/level-7-use-custom-jinja2-filter-test-n-global/custom-jj2-plugin/global.py
@@ -1,3 +1,3 @@
-from moban.jinja2.extensions import jinja_global
+from moban.plugins.jinja2.extensions import jinja_global
 
 jinja_global('global', dict(hello='world'))
