+++ moban/core/context.py
@@ -1,9 +1,11 @@
 import os
 import copy
 
-from moban import utils, reporter, constants, exceptions
+from moban import constants, exceptions
+from moban.core import utils
+from moban.externals import reporter
 from moban.program_options import OPTIONS
-from moban.data_loaders.manager import merge, load_data
+from moban.core.data_loader import merge, load_data
 
 
 class Context(object):
