+++ moban/deprecated/__init__.py
@@ -1,7 +1,9 @@
 import sys
 from functools import wraps
 
-from moban import plugins, reporter, constants, file_system
+from moban import constants
+from moban.core import plugins
+from moban.externals import reporter, file_system
 from moban.deprecated import repo
 from moban.deprecated.repo import git_clone
 from moban.deprecated.library import LIBRARIES
