+++ moban/plugins/jinja2/filters/github.py
@@ -1,6 +1,6 @@
 import re
 
-from moban.jinja2.extensions import JinjaFilter
+from moban.plugins.jinja2.extensions import JinjaFilter
 
 GITHUB_REF_PATTERN = "`([^`]*?#[0-9]+)`"
 ISSUE = "^.*?" + GITHUB_REF_PATTERN + ".*?$"
