+++ moban/plugins/jinja2/filters/repr.py
@@ -1,4 +1,4 @@
-from moban.jinja2.extensions import JinjaFilter
+from moban.plugins.jinja2.extensions import JinjaFilter
 
 
 @JinjaFilter()
