+++ moban/plugins/json_loader.py
@@ -2,7 +2,7 @@ import json
 
 from moban import constants
 from lml.plugin import PluginInfo
-from moban.file_system import open_file
+from moban.externals.file_system import open_file
 
 
 @PluginInfo(constants.DATA_LOADER_EXTENSION, tags=["json"])
