+++ tests/mobanfile/test_targets.py
@@ -4,7 +4,7 @@ import fs.path
 from nose.tools import eq_, raises
 from moban.mobanfile import targets
 from moban.exceptions import GroupTargetNotFound
-from moban.definitions import TemplateTarget
+from moban.core.definitions import TemplateTarget
 
 TEMPLATE = "a.jj2"
 OUTPUT = "a.output"
