+++ tests/mobanfile/test_templates.py
@@ -15,7 +15,7 @@ class TestHandleTemplateFunction:
         expected = [("copier-test01.csv", "/tmp/test", "csv")]
         eq_(expected, results)
 
-    @patch("moban.reporter.report_error_message")
+    @patch("moban.externals.reporter.report_error_message")
     def test_file_not_found(self, reporter):
         list(
             handle_template(
@@ -64,7 +64,7 @@ class TestHandleTemplateFunction:
             sorted(expected, key=lambda x: x[0]),
         )
 
-    @patch("moban.reporter.report_error_message")
+    @patch("moban.externals.reporter.report_error_message")
     def test_listing_dir_recusively_with_error(self, reporter):
         test_dir = "/tmp/copy-a-directory"
         list(
