+++ tests/test_hash_store.py
@@ -2,8 +2,8 @@ import os
 import sys
 
 from nose import SkipTest
-from moban import file_system
-from moban.hashstore import HashStore
+from moban.externals import file_system
+from moban.core.hashstore import HashStore
 
 
 class TestHashStore:
