+++ tests/test_reporter.py
@@ -1,8 +1,8 @@
 import sys
 
-import moban.reporter as reporter
 from mock import patch
 from nose.tools import eq_
+from moban.externals import reporter
 
 PY2 = sys.version_info[0] == 2
 if PY2:
